+++ applications/fishing-map/src/features/analysis/Analysis.tsx
@@ -1,36 +1,28 @@
-import { useEffect, useRef, useMemo, useState, Fragment } from 'react'
+import { useEffect, useRef, useState, Fragment } from 'react'
 import cx from 'classnames'
 import { event as uaEvent } from 'react-ga'
 import { useTranslation } from 'react-i18next'
-import union from '@turf/union'
-import { Feature, Polygon } from 'geojson'
 import { checkExistPermissionInList } from 'auth-middleware/src/utils'
 import { batch, useDispatch, useSelector } from 'react-redux'
 import { DateTime } from 'luxon'
-import bbox from '@turf/bbox'
 import GFWAPI from '@globalfishingwatch/api-client'
 import { Button, Icon, IconButton, Spinner } from '@globalfishingwatch/ui-components'
 import { Dataset, DatasetTypes } from '@globalfishingwatch/api-types'
 import { useFeatureState } from '@globalfishingwatch/react-hooks/dist/use-map-interaction'
-import { DEFAULT_CONTEXT_SOURCE_LAYER } from '@globalfishingwatch/layer-composer/dist/generators'
 import { useLocationConnect } from 'routes/routes.hook'
 import sectionStyles from 'features/workspace/shared/Sections.module.css'
 import { selectWorkspaceStatus } from 'features/workspace/workspace.selectors'
 import { selectUserData } from 'features/user/user.slice'
 import { AsyncReducerStatus } from 'utils/async-slice'
-import { selectAnalysisQuery } from 'features/app/app.selectors'
 import useMapInstance from 'features/map/map-context.hooks'
-import { useMapFitBounds } from 'features/map/map-viewport.hooks'
-import { useFeatures } from 'features/map/map-features.hooks'
-import { Bbox } from 'types'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
 import {
-  isGuestUser,
   selectActiveActivityDataviews,
   selectHasAnalysisLayersVisible,
 } from 'features/dataviews/dataviews.selectors'
 import { getRelatedDatasetByType } from 'features/datasets/datasets.selectors'
 import { getActivityFilters, getEventLabel } from 'utils/analytics'
+import { isGuestUser } from 'features/user/user.selectors'
 import styles from './Analysis.module.css'
 import {
   clearAnalysisGeometry,
@@ -39,47 +31,43 @@ import {
   DateRange,
   ReportGeometry,
   resetReportStatus,
-  selectAnalysisBounds,
   selectAnalysisAreaName,
   selectAnalysisGeometry,
   selectReportStatus,
-  setAnalysisGeometry,
 } from './analysis.slice'
 import AnalysisItem from './AnalysisItem'
-import { useFilteredTimeSeries } from './analysis.hooks'
-
-// const { filterByPolygon } = createAnalysisWorker<typeof AnalysisWorker>()
+import { useAnalysisGeometry, useFilteredTimeSeries } from './analysis.hooks'
 
 function Analysis() {
   const { t } = useTranslation()
   const workspaceStatus = useSelector(selectWorkspaceStatus)
   const { start, end } = useTimerangeConnect()
   const dispatch = useDispatch()
-  const fitMapBounds = useMapFitBounds()
-  const analysisQuery = useSelector(selectAnalysisQuery)
   const timeoutRef = useRef<NodeJS.Timeout>()
   const { dispatchQueryParams } = useLocationConnect()
-  const { updateFeatureState, cleanFeatureState } = useFeatureState(useMapInstance())
+  const { cleanFeatureState } = useFeatureState(useMapInstance())
   const { timerange } = useTimerangeConnect()
   const dataviews = useSelector(selectActiveActivityDataviews) || []
   const analysisGeometry = useSelector(selectAnalysisGeometry)
-  const analysisBounds = useSelector(selectAnalysisBounds)
+  const userData = useSelector(selectUserData)
   const guestUser = useSelector(isGuestUser)
 
   const analysisAreaName = useSelector(selectAnalysisAreaName)
   const reportStatus = useSelector(selectReportStatus)
-  const userData = useSelector(selectUserData)
   const hasAnalysisLayers = useSelector(selectHasAnalysisLayersVisible)
-  const { areaId, sourceId: contextSourceId } = analysisQuery
-  const contextSourcesIds = useMemo(() => [contextSourceId], [contextSourceId])
-  const filter = useMemo(() => ['==', 'gfw_id', parseInt(areaId)], [areaId])
 
-  const { sourcesFeatures: contextAreaFeaturesArr, haveSourcesLoaded: contextAreaSourceLoaded } =
-    useFeatures({
-      sourcesIds: contextSourcesIds,
-      sourceLayer: DEFAULT_CONTEXT_SOURCE_LAYER,
-      filter,
-    })
+  const datasetsReportAllowed = dataviews.flatMap((dataview) => {
+    const datasets: Dataset[] = (dataview?.config?.datasets || [])
+      .map((id: string) => dataview.datasets?.find((dataset) => dataset.id === id))
+      .map((dataset: Dataset) => getRelatedDatasetByType(dataset, DatasetTypes.Tracks))
+      .filter((dataset: Dataset) => {
+        const permission = { type: 'dataset', value: dataset.id, action: 'report' }
+        return checkExistPermissionInList(userData?.permissions, permission)
+      })
+    return datasets
+  })
+  const AISInDatasetsReport =
+    datasetsReportAllowed.find((dataset) => dataset.id.includes('global')) !== undefined
 
   const [timeRangeTooLong, setTimeRangeTooLong] = useState<boolean>(true)
 
@@ -91,73 +79,6 @@ function Analysis() {
       setTimeRangeTooLong(duration.years > 1)
     }
   }, [start, end])
-
-  useEffect(() => {
-    if (contextAreaSourceLoaded) {
-      cleanFeatureState('highlight')
-      const featureState = {
-        source: contextSourceId,
-        sourceLayer: DEFAULT_CONTEXT_SOURCE_LAYER,
-        id: areaId,
-      }
-      updateFeatureState([featureState], 'highlight')
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [areaId, contextSourceId, contextAreaSourceLoaded])
-
-  useEffect(() => {
-    if (contextAreaSourceLoaded) {
-      const contextAreaFeatures = contextAreaFeaturesArr?.[0]
-      if (contextAreaFeatures && contextAreaFeatures.length > 0) {
-        const contextAreaFeatureMerged = contextAreaFeatures.reduce(
-          (acc, { geometry, properties }) => {
-            const featureGeometry: Feature<Polygon> = {
-              type: 'Feature',
-              geometry: geometry as Polygon,
-              properties,
-            }
-            if (!acc?.type) return featureGeometry
-            return union(acc, featureGeometry, { properties }) as Feature<Polygon>
-          },
-          {} as Feature<Polygon>
-        )
-        const { name, value, id } = contextAreaFeatureMerged.properties || {}
-        const areaName = name || id || value
-        if (areaName !== analysisAreaName) {
-          const bounds = bbox(contextAreaFeatureMerged) as Bbox
-          dispatch(
-            setAnalysisGeometry({
-              geometry: contextAreaFeatureMerged,
-              name: name || id || value,
-              bounds,
-            })
-          )
-        }
-      } else {
-        console.warn('No feature for analysis found')
-        dispatch(
-          setAnalysisGeometry({
-            geometry: undefined,
-            name: 'Not found 🙈',
-          })
-        )
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [contextAreaFeaturesArr, dispatch, contextAreaSourceLoaded])
-
-  useEffect(() => {
-    if (analysisBounds) {
-      fitMapBounds(analysisBounds, { padding: 10 })
-      dispatchQueryParams({
-        analysis: {
-          ...analysisQuery,
-          bounds: analysisBounds,
-        },
-      })
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [analysisBounds])
 
   const onCloseClick = () => {
     cleanFeatureState('highlight')
@@ -216,15 +137,25 @@ function Analysis() {
     }
   }, [])
 
-  const { generatingTimeseries, haveSourcesLoaded, sourcesTimeseriesFiltered } =
-    useFilteredTimeSeries()
+  const layersTimeseriesFiltered = useFilteredTimeSeries()
+  const analysisGeometryLoaded = useAnalysisGeometry()
+
+  let downloadTooltip = ''
+  if (timeRangeTooLong) {
+    downloadTooltip = t(
+      'analysis.timeRangeTooLong',
+      'Reports are only allowed for time ranges up to a year'
+    )
+  } else if (!AISInDatasetsReport) {
+    downloadTooltip = t('analysis.onlyAISAllowed', 'Only AIS datasets are allowed to download')
+  }
 
   return (
     <div className={styles.container}>
       <div className={styles.header}>
         <h2 className={styles.sectionTitle}>
           {t('analysis.title', 'Analysis')}
-          <span className={styles.error}> Experimental</span>
+          <span className={styles.experimental}>{t('analysis.experimental', 'Experimental')}</span>
         </h2>
         <div className={cx('print-hidden', sectionStyles.sectionButtons)}>
           <IconButton
@@ -237,32 +168,31 @@ function Analysis() {
         </div>
       </div>
       {workspaceStatus !== AsyncReducerStatus.Finished ||
-      !contextAreaSourceLoaded ||
-      !analysisGeometry ||
-      !haveSourcesLoaded ||
-      generatingTimeseries ? (
+      !analysisGeometryLoaded ||
+      !layersTimeseriesFiltered ? (
         <Spinner className={styles.spinnerFull} />
       ) : (
         <div className={styles.contentContainer}>
           <div className={styles.content}>
-            {sourcesTimeseriesFiltered && sourcesTimeseriesFiltered?.length ? (
+            {layersTimeseriesFiltered && layersTimeseriesFiltered?.length ? (
               <Fragment>
-                {sourcesTimeseriesFiltered.map((sourceTimeseriesFiltered, index) => {
+                {layersTimeseriesFiltered.map((layerTimeseriesFiltered, index) => {
                   return (
                     <AnalysisItem
                       hasAnalysisLayers={hasAnalysisLayers}
                       analysisAreaName={analysisAreaName}
                       key={index}
-                      graphData={sourceTimeseriesFiltered}
+                      graphData={layerTimeseriesFiltered}
                     />
                   )
                 })}
               </Fragment>
             ) : (
-              <p className={styles.emptyDataPlaceholder}>No data available</p>
+              <p className={styles.emptyDataPlaceholder}>
+                {t('analysis.noData', 'No data available')}
+              </p>
             )}
-
-            {contextAreaSourceLoaded && analysisGeometry && (
+            {analysisGeometry && (
               <p className={styles.placeholder}>
                 {t(
                   'analysis.disclaimer',
@@ -303,18 +233,12 @@ function Analysis() {
                   className={styles.saveBtn}
                   onClick={onDownloadClick}
                   loading={reportStatus === AsyncReducerStatus.LoadingCreate}
-                  tooltip={
-                    timeRangeTooLong
-                      ? t(
-                          'analysis.timeRangeTooLong',
-                          'Reports are only allowed for time ranges up to a year'
-                        )
-                      : ''
-                  }
+                  tooltip={downloadTooltip}
                   tooltipPlacement="top"
                   disabled={
                     timeRangeTooLong ||
                     !hasAnalysisLayers ||
+                    !AISInDatasetsReport ||
                     reportStatus === AsyncReducerStatus.Finished
                   }
                 >