+++ applications/fishing-map/src/features/map/map-features.hooks.ts
@@ -1,11 +1,7 @@
-import { useEffect, useMemo } from 'react'
+import { useEffect } from 'react'
 import { atom, useRecoilValue, useSetRecoilState } from 'recoil'
-import { useSelector } from 'react-redux'
-import { TEMPORALGRID_SOURCE_LAYER } from '@globalfishingwatch/layer-composer'
-import { AnyGeneratorConfig } from '@globalfishingwatch/layer-composer/dist/generators/types'
 import useMapInstance from 'features/map/map-context.hooks'
 import { useMapStyle } from './map.hooks'
-import { selectActiveHeatmapAnimatedGeneratorConfigs } from './map.selectors'
 
 export const mapIdleAtom = atom<boolean>({
   key: 'mapIdle',
@@ -54,81 +50,15 @@ export const useMapLoaded = () => {
   return loaded
 }
 
-export const useHaveSourcesLoaded = (sourcesIds: string | string[]) => {
+export const useSourceInStyle = (sourcesIds: string | string[]) => {
+  const style = useMapStyle()
   const sourcesIdsList = Array.isArray(sourcesIds) ? sourcesIds : [sourcesIds]
-  const mapLoaded = useMapLoaded()
-  const style = useMapStyle()
-
   const sourcesLoaded = sourcesIdsList.every((source) => style?.sources?.[source] !== undefined)
+  return sourcesLoaded
+}
+
+export const useMapAndSourcesLoaded = (sourcesIds: string | string[]) => {
+  const mapLoaded = useMapLoaded()
+  const sourcesLoaded = useSourceInStyle(sourcesIds)
   return mapLoaded && sourcesLoaded
-}
-
-export const useActiveHeatmapAnimatedMetadatas = (generators: AnyGeneratorConfig[]) => {
-  const style = useMapStyle()
-
-  const generatorsIds = generators?.map((generator) => generator.id)
-  const generatorsMetadata = generatorsIds?.flatMap((generatorId) => {
-    return style?.metadata?.generatorsMetadata?.[generatorId] || []
-  })
-
-  const serializedGeneratorIds =
-    generatorsMetadata &&
-    generatorsMetadata
-      .map((metadata) => {
-        return metadata?.timeChunks?.activeSourceId + metadata?.numSublayers
-      })
-      .join()
-
-  const metadatas = useMemo(() => {
-    return generatorsMetadata || []
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [serializedGeneratorIds])
-  return metadatas
-}
-
-export const useFeatures = ({
-  sourcesIds,
-  sourceLayer = 'main',
-  filter,
-}: {
-  sourcesIds: string[]
-  sourceLayer?: string
-  filter?: any[]
-}) => {
-  const haveSourcesLoaded = useHaveSourcesLoaded(sourcesIds)
-  const map = useMapInstance()
-
-  const sourcesFeatures = useMemo(() => {
-    if (haveSourcesLoaded && map) {
-      const features = sourcesIds.map((sourceId) => {
-        const sourceFeatures = map.querySourceFeatures(sourceId, {
-          sourceLayer,
-          ...(filter && { filter }),
-        })
-        return sourceFeatures
-      })
-      return features
-    }
-  }, [haveSourcesLoaded, map, sourceLayer, filter, sourcesIds])
-
-  return { sourcesFeatures, haveSourcesLoaded }
-}
-
-const useGeneratorAnimatedFeatures = (generators: AnyGeneratorConfig[]) => {
-  const sourcesMetadata = useActiveHeatmapAnimatedMetadatas(generators)
-
-  const sourcesIds: string[] = useMemo(() => {
-    return sourcesMetadata?.map((metadata) => metadata?.timeChunks?.activeSourceId)
-  }, [sourcesMetadata])
-
-  const { sourcesFeatures, haveSourcesLoaded } = useFeatures({
-    sourcesIds,
-    sourceLayer: TEMPORALGRID_SOURCE_LAYER,
-  })
-  return { sourcesFeatures, haveSourcesLoaded, sourcesMetadata }
-}
-
-export const useActiveHeatmapAnimatedFeatures = () => {
-  const generators = useSelector(selectActiveHeatmapAnimatedGeneratorConfigs)
-  return useGeneratorAnimatedFeatures(generators)
 }