+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @globalfishingwatch/react-hooks
 
+## 7.0.3
+
+### Patch Changes
+
+- Updated dependencies [9e507541]
+  - @globalfishingwatch/layer-composer@7.0.3
+
+## 7.0.2
+
+### Patch Changes
+
+- Updated dependencies [b230075f]
+  - @globalfishingwatch/fourwings-aggregate@4.0.2
+  - @globalfishingwatch/layer-composer@7.0.2
+
 ## 7.0.1
 
 ### Patch Changes